Revert back to individual native implementations of SecureRandom instead of the native methods (Linux will need to hold onto some internal references
