--- secure-random/src/darwinMain/kotlin/io/matthewnelson/component/secure/random/internal/-DarwinPlatform.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2023 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- **/
-package io.matthewnelson.component.secure.random.internal
-
-import kotlinx.cinterop.UnsafeNumber
-import kotlinx.cinterop.addressOf
-import kotlinx.cinterop.convert
-import kotlinx.cinterop.usePinned
-import platform.Security.SecRandomCopyBytes
-import platform.Security.kSecRandomDefault
-
-@OptIn(UnsafeNumber::class)
-internal actual fun ByteArray.nativeNextBytes() {
-    val size = size.toUInt()
-    usePinned { pinned ->
-        SecRandomCopyBytes(kSecRandomDefault, size.convert(), pinned.addressOf(0))
-    }
-}
