+++ secure-random/src/linuxMain/kotlin/io/matthewnelson/component/secure/random/SecureRandom.kt
@@ -12,6 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  **/
 package io.matthewnelson.component.secure.random
 
@@ -20,13 +21,8 @@ import io.matthewnelson.component.secure.random.internal.commonNextBytesOf
 
 /**
  * A cryptographically strong random number generator (RNG).
- *
- * @see [java.security.SecureRandom]
  * */
-public actual class SecureRandom: java.security.SecureRandom {
-
-    public actual constructor(): super()
-    public constructor(seed: ByteArray): super(seed)
+public actual class SecureRandom public actual constructor() {
 
     /**
      * Returns a [ByteArray] of size [count], filled with
@@ -39,12 +35,13 @@ public actual class SecureRandom: java.security.SecureRandom {
 
     /**
      * Fills a [ByteArray] with securely generated random data.
-     *
-     * @see [java.security.SecureRandom.nextBytes]
      * */
-    public actual override fun nextBytes(bytes: ByteArray?) {
+    public actual fun nextBytes(bytes: ByteArray?) {
         bytes.ifNotNullOrEmpty {
-            super.nextBytes(this)
+            // TODO
+            for (i in this.indices) {
+                this[i] = 1
+            }
         }
     }
 }
